Nove difetti trovati da una caccia strutturata, il primo era mio di stamattina
Venticinque agenti su cinque lenti (posti e turni, conti scritti due volte,
eccezioni non gestite, flussi senza uscita, dati che il database rifiuta), ogni
sospetto verificato da un secondo agente col compito di smontarlo: dodici
confermati, quattro respinti.

1. IL TORNEO ERA BLOCCATO TRE MANI SU QUATTRO. Aggiungendo stamattina il blocco
   delle dichiarazioni fuori turno ho creato, nella pagina nuova del torneo, una
   mano che non parte: col mazziere diverso da Sud il cassetto non compare — non
   è il tuo turno — e nessuno faceva parlare i robot. Nessun pulsante, nessuna
   via d'uscita, e il bottone «Prossima mano» sta dentro il blocco del voto che
   non arriva mai. La pagina gemella lo faceva già, col commento che spiegava
   perché: è esattamente il difetto che una prova E2E del torneo avrebbe preso,
   e che non esisteva. Ora esiste.

2. LA LICITA CON UN AMICO POTEVA MORIRE: timeout di 15 secondi verso BEN senza
   `maxDuration`, cioè più lungo del tetto della funzione. Stesso difetto già
   corretto sulle route /api/ben/*, rimasto qui.

3. IL WORKER DDS CHE MUORE CONGELAVA LA PARTITA: `pending.clear()` senza
   rispondere lasciava le promesse appese per sempre, e la rete di sicurezza col
   timeout non scattava perché cercava una richiesta che era stata cancellata.
   «Verifica del reclamo…» eterno.

4. IL DATABASE RIFIUTAVA IN SILENZIO. `profile_type` non ammetteva «junior»:
   chi sceglie quel profilo faceva fallire l'UPDATE che porta con sé XP,
   striscia e minuti — un ragazzino che gioca non vedeva salire niente, senza un
   messaggio. E `game_type` non conosceva «segnali» né «sfida-settimanale», che
   restavano in coda offline a riprovare all'infinito.

5. LA SCATOLA DEL RIPASSO NON ESISTEVA nel database: chi cambiava dispositivo
   ritrovava ogni domanda nella prima scatola e rifaceva un ripasso già fatto.
   Il metodo di Leitner c'era da mesi e non funzionava.

6. LA ROTELLA CHE GIRA PER SEMPRE, in otto store: il `catch` non segnava lo
   stato come terminale, quindi l'effetto che carica ripartiva subito — la
   guardia è `!isLoaded && !isLoading` e il catch metteva isLoading a false.
   Richieste a raffica e pagina che non arriva mai.

7. LA PROMESSA AVVELENATA: il modulo WebAssembly rifiutato restava in cache per
   tutta la sessione, quindi un «Riprova» non poteva riuscire per definizione.

8. IL CONTRO PRESO PER CONTRATTO, terza copia. `contrattoFinale` in
   licita-a-due mostrava «Contratto: X» invece di «1♠ contrato». Le altre due
   copie erano già state corrette, a mesi di distanza l'una dall'altra.

9. IL TAVOLO DI STUDIO VUOTO: con un link vecchio (`?mano=`) la mano non si
   caricava, il calcolo del contratto veniva saltato perché il parametro c'era,
   e restava un tavolo morto senza spiegazione.

Restano da fare le ancore di rotazione in /gioca/smazzata (la lettera «S»
indica due giocatori diversi fra tavolo e griglia d'asta): è in quattro file e
va fatto con calma.

Verifiche: 992 test, tsc, eslint, build, test:rls, prova-campo,
e2e/riepilogo + e2e/torneo-licita.

+++ scripts/sql/scenari-e-mani-2026-08.sql
@@ -278,7 +278,11 @@ as $function$
       WHERE r.mano_id = p_mano_id AND r.user_id <> auth.uid()
     ),
     'contratti', (
-      SELECT coalesce(jsonb_agg(x ORDER BY x->>'quanti' DESC), '[]'::jsonb)
+      -- `->>` restituisce testo: ordinando così, «9» verrebbe prima di «10».
+      -- Il database in produzione ha già la versione giusta; questo file è la
+      -- fonte da cui si ricostruisce tutto, e una differenza qui è una bomba a
+      -- orologeria per il giorno in cui lo si rieseguirà.
+      SELECT coalesce(jsonb_agg(x ORDER BY (x->>'quanti')::int DESC), '[]'::jsonb)
       FROM (
         SELECT jsonb_build_object(
           'contratto', coalesce(r.contratto, 'passo'),
